19_lesson/requirements.txt
fastapi
fastapi-users[sqlalchemy]
uvicorn[standard]
asyncpg
Pillow
python-dotenv